+++ common/src/main/java/cc/cassian/pyrite/config/ModConfig.java
@@ -1,0 +1,16 @@
+package cc.cassian.pyrite.config;
+
+import folk.sisby.kaleido.api.WrappedConfig;
+import folk.sisby.kaleido.lib.quiltconfig.api.annotations.Comment;
+
+public class ModConfig extends WrappedConfig {
+   @Comment("Enable the Oddities - removed features from Pyrite")
+   public boolean oddities = true;
+   @Comment("Enable Pyrite Azalea Trees and their wood set.")
+   public boolean azalea = true;
+   @Comment("Enable Pyrite Tall Mushrooms and their wood set.")
+   public boolean mushrooms = true;
+   @Comment("Add a tooltip on disabled items.")
+   public boolean disabledContentTooltip = true;
+
+}